+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.0 - 2018-12-07
 
 * Add `SlaStatusInput`
+* Fix tab on SelectInput
 
 ## 0.6.9 - 2018-12-06